pyproject.toml
[tool.poetry]
name          = "onboardme"
version       = "1.7.0"
description   = "Install dot files and packages, including a base mode with sensible defaults to run on most computers running Debian based distros or macOS."
authors       = [
    "Jesse Hitch <jessebot@linux.com>",
    "Max Roby <emax@cloudydev.net>"
]
license       = "AGPL-3.0-or-later"
readme        = "README.md"
documentation = "https://jessebot.github.io/onboardme/onboardme"
repository    = "http://github.com/jessebot/onboardme"
keywords      = ["onboardme", "onboarding", "desktop-setup", "development-environment"]
classifiers   = ["Development Status :: 3 - Alpha",
                 "Programming Language :: Python :: 3.11",
                 "Operating System :: MacOS :: MacOS X",
                 "Operating System :: POSIX :: Linux",
                 "Intended Audience :: End Users/Desktop",
                 "Topic :: System :: Installation/Setup",
                 "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)"]
packages      = [{include = "onboardme"}]
include       = ["onboardme/scripts/update_apt_sources.sh"]

[tool.poetry.dependencies]
python              = "^3.11"
click               = "^8.1"
rich                = "^13.6"
PyYAML              = "^6.0"
GitPython           = "^3.1"
wget                = "^3.2"
xdg-base-dirs       = "^6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5"

[tool.poetry.plugins."onboardme.application.plugin"]
"onboardme" = "onboardme:main"

[tool.poetry.scripts]
onboardme = 'onboardme:main'

[tool.poetry.urls]
"Bug Tracker" = "http://github.com/jessebot/onboardme/issues"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
